test(core): split react refresh tests
